Extract config validation class into separate module
