+++ pulsar-config-validation/src/main/java/org/apache/pulsar/config/validation/ConfigValidationUtils.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pulsar.common.validator;
+package org.apache.pulsar.config.validation;
 
 import java.util.Map;
